Made slight refactors to the code base.
