+++ src/components/common/SectionInfo/index.ts
@@ -1,0 +1,1 @@
+export { default as SectionInfo } from './SectionInfoContainer';
