+++ src/components/common/index.ts
@@ -12,6 +12,7 @@ export { LogoutPopup } from './LogoutPopup';
 export { Modal } from './Modal';
 export { PrivateRoute } from './PrivateRoute';
 export { PromptBox } from './PromptBox';
+export { SectionInfo } from './SectionInfo';
 export * as Select from './Select';
 export { SEO } from './SEO';
 export * as Table from './Table';
