--- src/components/pages/TeacherDashboard/TeacherSectionInfo/index.ts
@@ -1,1 +1,0 @@
-export { default as TeacherSectionInfo } from './TeacherSectionInfoContainer';
