+++ src/styles/components/pages/teacherDashboard/teacherDashboard.scss
@@ -13,4 +13,3 @@
 @import './teacherViewSectionPage/teacherViewSection.scss';
 @import './teacherViewRumblePage/teacherViewRumble.scss';
 @import './teacherDashboardStudentList.scss';
-@import './teacherSectionInfo.scss';
