+++ src/styles/index.scss
@@ -25,3 +25,4 @@
 @import './components/common/select.scss';
 @import './components/common/checkbboxGroup.scss';
 @import './components/common/loader.scss';
+@import './components/common/sectionInfo.scss';
